fix: extract test-guard helper, fix matrix rotate_key entries, toolInvocations aggregation, JSDoc, citations
Extract isTestEnvironment()/assertTestEnvironment() to internal/test-guard.ts
so the allowlist predicate is defined once; both assertTestHookAllowed (scp.ts)
and __setBridgeForTests (bridge.ts) delegate to it — closes security-relevant
guard drift.

Fix sdk-capability-matrix.json Identity/rotate_key: UniFFI bridge DOES export
rotate_key (ScpBindings.swift open func rotateKey()); correct exemption text for
swift and kotlin to accurately describe the gap (no public SDK wrapper, not a
missing bridge export).

Add comment to toolInvocations Layer 2 matching Python SDK (both are thin
facades; no aggregation by tool ID in either implementation).

Remove non-existent top-level ok from economyVerifyPaymentReceipts JSDoc;
remove unreachable @throws EconomicPolicyUnsupportedOnWasm (SCP class is
NAPI-only).

Fix identityMigrate spec citation in scp.ts and scp.py: §3.2.1 (same-DID
custody migration) -> §3/ADR-003 §4b (new-DID migration + DidRotationEvent).

Remove stale renew_attestation TODO comment from ALIASES table in
check-sdk-coverage.py.

Add _node_text() helper replacing all 23 (node.text or b'').decode('utf-8')
repetitions in the SDK coverage gate script.
